feat: add setting to delete pin message notifs

+++ src/monitors/pinsAdd.ts
@@ -1,6 +1,6 @@
 import { Monitor, MonitorOptions, KlasaMessage } from 'klasa';
-import { Message } from 'discord.js';
 import { ApplyOptions } from '@skyra/decorators';
+import { GuildSettings } from '@lib/types/settings/GuildSettings';
 
 @ApplyOptions<MonitorOptions>({
 	allowedTypes: ['PINS_ADD'],
@@ -8,8 +8,11 @@ import { ApplyOptions } from '@skyra/decorators';
 })
 export default class extends Monitor {
 
-	public async run(msg: KlasaMessage): Promise<Message> {
-		return msg.delete();
+	public async run(msg: KlasaMessage) {
+		if (msg.guild) {
+			const deletePinMessages = msg.guild!.settings.get(GuildSettings.DeletePinMessages) as boolean;
+			if (deletePinMessages) return msg.delete();
+		}
 	}
 
 }
